BE/requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.7.1
psycopg2-binary==2.9.10
Flask-Cors==5.0.0